backend/requirements.txt
anyio==3.6.2
Authlib==1.1.0
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.85.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
httpx-oauth==0.10.1
idna==3.4
itsdangerous==2.1.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
termcolor==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0

